shared pass for multiple hardware
