+++ paddle/fluid/inference/api/paddle_pass_builder.cc
@@ -614,6 +614,7 @@ const std::vector<std::string> kPirGpuPasses{
     "matmul_add_act_fuse_pass",
     "fc_elementwise_layernorm_fuse_pass",
     "add_norm_fuse_pass",
+    "group_norm_silu_fuse_pass",
     "matmul_scale_fuse_pass",
     "matmul_transpose_fuse_pass",
     "transpose_flatten_concat_fuse_pass",
@@ -627,7 +628,7 @@ const std::vector<std::string> kPirXpuPasses{// Functional pass
                                              // Operator fusion pass
                                              "add_layernorm_xpu_fuse_pass",
                                              "conv2d_bn_xpu_fuse_pass",
-                                             "group_norm_silu_xpu_fuse_pass"};
+                                             "group_norm_silu_fuse_pass"};
 
 const std::vector<std::string> kPirMkldnnPasses {
   "depthwise_conv_onednn_pass",              //
