+++ paddle/fluid/pir/transforms/general/group_norm_silu_fuse_pass.h
@@ -21,6 +21,6 @@ namespace pir {
 
 class Pass;
 
-IR_API std::unique_ptr<Pass> CreateGroupNormSiluXpuFusePass();
+IR_API std::unique_ptr<Pass> CreateGroupNormSiluFusePass();
 
 }  // namespace pir
