+++ test/ir/pir/fused_pass/test_add_norm_fuse_pass.py
@@ -552,93 +552,5 @@ class TestAddGroupNormPatternSilu_FP16(PassTest):
         self.check_pass_correct()
 
 
-class GroupNormSiluPattern_FP16(PassTest):
-    r"""
-    group_norm
-        |
-      silu
-    """
-
-    def is_program_valid(self, program=None):
-        return True
-
-    def sample_program(self):
-        for x_shape in [[2, 6, 4, 2]]:
-            for residual_shape in [[1, 6, 1, 1]]:
-                for dtype in ['float16']:
-                    for epilson in [1e-5]:
-                        for groups in [2]:
-                            for data_layout in ['NCHW']:
-                                rand_value = (
-                                    0.001
-                                    * paddle.rand(
-                                        shape=[x_shape[1]], dtype=dtype
-                                    ).numpy()
-                                )
-                                with paddle.pir_utils.IrGuard():
-                                    start_prog = paddle.static.Program()
-                                    main_prog = paddle.static.Program()
-                                    with paddle.pir.core.program_guard(
-                                        main_prog, start_prog
-                                    ):
-                                        x = paddle.static.data(
-                                            name='x', shape=x_shape, dtype=dtype
-                                        )
-                                        w = create_parameter(
-                                            shape=[x_shape[1]],
-                                            dtype=dtype,
-                                            initializer=paddle.nn.initializer.Assign(
-                                                rand_value
-                                            ),
-                                        )
-                                        b = create_parameter(
-                                            shape=[x_shape[1]],
-                                            dtype=dtype,
-                                            initializer=paddle.nn.initializer.Assign(
-                                                rand_value
-                                            ),
-                                        )
-                                        group_norm_out = (
-                                            paddle.nn.functional.group_norm(
-                                                x,
-                                                num_groups=groups,
-                                                epsilon=epilson,
-                                                weight=w,
-                                                bias=b,
-                                                data_format=data_layout,
-                                            )
-                                        )
-                                        out = paddle.nn.functional.silu(
-                                            group_norm_out
-                                        )
-                                        out = paddle.assign(out)
-                                        self.pass_attr_list = [
-                                            {'add_norm_fuse_pass': {}},
-                                            {'transfer_layout_pass': {}},
-                                            {
-                                                'remove_redundant_transpose_pass': {}
-                                            },
-                                        ]
-                                        self.feeds = {
-                                            "x": np.random.random(
-                                                x_shape
-                                            ).astype(dtype),
-                                        }
-                                        self.fetch_list = [out]
-                                        self.valid_op_map = {
-                                            "pd_op.silu": 0,
-                                            "pd_op.group_norm": 0,
-                                            "pd_op.add_group_norm_silu": 1,
-                                        }
-                                        yield [main_prog, start_prog], False
-
-    def setUp(self):
-        if core.is_compiled_with_cuda():
-            self.places.append(paddle.CUDAPlace(0))
-
-    def test_check_output(self):
-        self.check_pass_correct()
-
-
 if __name__ == "__main__":
     unittest.main()
